pyproject.toml
[tool.poetry]
name = "wms"
version = "0.1.0"
description = "A web-native and LLM-enhanced app for organizing and finding your sh*t!"
authors = ["Alex Falconer-Athanassakos"]
readme = "README.md"
package-mode = false
[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.3"
psycopg2-binary = "^2.9.10"
pillow = ">=11.3.0"
pillow-heif = "^1.0.0"
qrcode = "^8.0"
langchain = "0.3.27"
langchain-aws = "0.2.29"
ruff = "^0.11.5"
gunicorn = "^22.0.0"
whitenoise = "^6.7.0"
django-storages = {extras = ["boto3"], version = "^1.14.6"}
pydantic = "^2.12.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "wms.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
markers = [
    "e2e: End-to-end browser tests using Browser-Use (deselect with '-m \"not e2e\"')",
]
asyncio_mode = "auto"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D100", # Docstring for modules
    "D105", # Docstring for magic methods
    "D413", # Blank line at end of docstring
    "D204", # Blank line after docstring
    "COM812", # Trailing comma in function def
    "FBT001", # Boolean-typed positional argument in function definition
    "FBT002", # Boolean default positional argument in function definition
]


[tool.ruff]
line-length = 150

[tool.ruff.per-file-ignores]
"{test_*.py,conftest.py}" = [
    "D100",  # Missing docstring in public module
    "D101",  # Missing docstring in public class
    "D102",  # Missing docstring in public method
    "D103",  # Missing docstring in public function
    "D104",  # Missing docstring in public package
    "D105",  # Missing docstring in magic method
    "S101", # Use of assert
    "S311", # Standard pseudo-random generators are not suitable for cryptographic purposes
    "ANN201",
    "S106",  # Possible hardcoded password assigned to argument: "password"
    "PLR2004"  # Magic value used in comparison, consider replacing `5.0` with a constant variable
]

"core/context_processors.py" = ["ARG001"]
[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
pytest-django = "^4.11.1"

[tool.poetry.group.e2e]
optional = true

[tool.poetry.group.e2e.dependencies]
browser-use = "^0.12.1"
pytest-asyncio = "^0.24.0"
